App commands now work
