+++ GrailDiary/DocumentListViewController.swift
@@ -190,6 +190,11 @@ final class DocumentListViewController: UIViewController {
     Logger.shared.info("Responder chain: \(responderChain())")
   }
 
+  override func viewDidAppear(_ animated: Bool) {
+    super.viewDidAppear(animated)
+    Logger.shared.info("Responder chain: \(responderChain())")
+  }
+
   override func viewWillDisappear(_ animated: Bool) {
     super.viewWillDisappear(animated)
     dataSource?.stopObservingDatabase()
